Move AssertUtils in internal
